Crate rinvaders

Main rust wasm module for the rinvaders game. This module contains the main Game struct, which holds all the game state and provides update, input and rendering functions for the game.

Modules

__rg_allocator_abi

entites
Module, which reexports all the private entity types.

input
Input handling module. This module takes care of all user input. It essentially transforms DOM keyboard / gamepad events into requested movement / action state for the player (in rust types).

logging
Easy logging module. Provides utilities for logging on the web platform.

rendering
Renderer module. Provides infrastructure to render the game scene to the canvas.

utils
Utilities module (various stuff).

Macros

log
Log macro. Takes a format string like the one taken by the rust formatter API, and a variadic number of arguments, which are interpolated into the resulting log message string.

Structs

Game
Main type that holds all game state and provides functionality to update and render the game.

Constants

TICK
Game tick. All 10 ms.

Statics

ALLOC

__WASM_BINDGEN_GENERATED_3efceea8801f90d8

__WASM_BINDGEN_GENERATED_5d6734b2387f80d7

__WASM_BINDGEN_GENERATED_e3c1a1a7c96ebeca

Functions

__wasm_bindgen_generated_Game_keyboard_event
Forwards keyboard events to the input module.

__wasm_bindgen_generated_Game_new
Create an instance of the Game type. Requires canvas parameter which should be the DOM node of the canvas, into which the game should be rendered.

__wasm_bindgen_generated_Game_render
Render the game. This will clear the canvas, and then redraw the whole scene. Does not need mutable access, because it immutably reads game date, and draws to the canvas. Canvas draw operations have some kind of interior mutability, which allows usage through non mutable references.

__wasm_bindgen_generated_Game_set_gamepad_state
Set the current gamepad state. Axis movement and shoot.

__wasm_bindgen_generated_Game_update
Update the game.

__wasm_bindgen_generated_hello
Test exported rust function (to wasm module).

__wbg_game_free⚠

__wbindgen_describe_game_keyboard_event

__wbindgen_describe_game_new

__wbindgen_describe_game_render

__wbindgen_describe_game_set_gamepad_state

__wbindgen_describe_game_update

__wbindgen_describe_hello

hello
Test exported rust function (to wasm module).