Module rinvaders::entites::enemy

Enemy entity module.

An enemy is one of those black things coming from the top of the screen. Their behaviour is simple:

Move down one pixel (virtual pixel, not a real pixel on the canvas context) every enemy tick.

When the void at the bottom is reached, die and remove some points from the player score.

When shot by projectile, die and give player score some points.

Spawn on random location on top of the screen every enemy spawn tick.

There are several visual types of enemies. They each have a different bounding box for collision.

Structs

ENEMY_PSO
Pso for enemies. Fill colored black pixels.

Enemy
The actual enemy entity. Contains position and enemy type.

Enums

EnemyType
Types of enemies (different visuals and bounding boxes)

Statics

ENEMY_PSO