Trait rinvaders::rendering::Renderable

pub trait Renderable {
    fn draw(&self, pxs: &PixelScreen);
}

Trait for renderable types.

Required Methods

fn draw(&self, pxs: &PixelScreen)

This will draw the current object to the given PixelScreen.

Note: PSO usage is not specified by Renderable. This method may rebind another or even multiple PSOs, but it does not have to.