Struct rinvaders::Game

pub struct Game {
    fps_counter: FpsCounter,
    renderer: PixelScreen,
    player: Player,
    input: Input,
    projectiles: Vec<Projectile>,
    enemies: Vec<Enemy>,
    game_tick: Timer,
    enemy_tick: u8,
    enemy_spawn_tick: u8,
    explosions: Vec<Explosion>,
    explosion_tick: u8,
    score: Score,
}

Main type that holds all game state and provides functionality to update and render the game.

Fields
fps_counter: FpsCounter renderer: PixelScreen player: Player input: Input projectiles: Vec<Projectile> enemies: Vec<Enemy> game_tick: Timer enemy_tick: u8 enemy_spawn_tick: u8 explosions: Vec<Explosion> explosion_tick: u8 score: Score
Methods

impl Game

pub fn new(canvas: HtmlCanvasElement) -> Self

Create an instance of the Game type. Requires canvas parameter which should be the DOM node of the canvas, into which the game should be rendered.

Initializes the game.

pub fn keyboard_event(&mut self, is_down: bool, e: KeyboardEvent)

Forwards keyboard events to the input module.

pub fn render(&self)

Render the game. This will clear the canvas, and then redraw the whole scene. Does not need mutable access, because it immutably reads game date, and draws to the canvas. Canvas draw operations have some kind of interior mutability, which allows usage through non mutable references.

pub fn update(&mut self, ts: u32)

Update the game.

This processes all input, and calculates the next state of the game depending on previous state, current input, and the current timestamp.

pub fn set_gamepad_state(&mut self, left: bool, right: bool, shoot: bool)

Set the current gamepad state. Axis movement and shoot.

This will forward the input to the input module.